update perf collectors

+++ internal/openmetrics-exporter/arrays_performance_collector.go
@@ -7,10 +7,10 @@ import (
 
 type ArrayPerformanceCollector struct {
 	LatencyDesc     *prometheus.Desc
-        ThroughputDesc  *prometheus.Desc
-        BandwidthDesc   *prometheus.Desc
-        AverageSizeDesc *prometheus.Desc
-        QueueDepthDesc  *prometheus.Desc
+	ThroughputDesc  *prometheus.Desc
+	BandwidthDesc   *prometheus.Desc
+	AverageSizeDesc *prometheus.Desc
+	QueueDepthDesc  *prometheus.Desc
 	Client          *client.FAClient
 }
 
@@ -27,198 +27,198 @@ func (c *ArrayPerformanceCollector) Collect(ch chan<- prometheus.Metric) {
 	ch <- prometheus.MustNewConstMetric(
 		c.LatencyDesc,
 		prometheus.GaugeValue,
-		float64(ap.QueueUsecPerMirroredWriteOp),
+		ap.QueueUsecPerMirroredWriteOp,
 		"queue_usec_per_mirrored_write_op",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.LatencyDesc,
 		prometheus.GaugeValue,
-		float64(ap.QueueUsecPerReadOp),
+		ap.QueueUsecPerReadOp,
 		"queue_usec_per_read_op",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.LatencyDesc,
 		prometheus.GaugeValue,
-		float64(ap.QueueUsecPerWriteOp),
+		ap.QueueUsecPerWriteOp,
 		"queue_usec_per_write_op",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.LatencyDesc,
 		prometheus.GaugeValue,
-		float64(ap.SanUsecPerMirroredWriteOp),
+		ap.SanUsecPerMirroredWriteOp,
 		"san_usec_per_mirrored_write_op",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.LatencyDesc,
 		prometheus.GaugeValue,
-		float64(ap.SanUsecPerReadOp),
+		ap.SanUsecPerReadOp,
 		"san_usec_per_read_op",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.LatencyDesc,
 		prometheus.GaugeValue,
-		float64(ap.SanUsecPerWriteOp),
+		ap.SanUsecPerWriteOp,
 		"san_usec_per_write_op",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.LatencyDesc,
 		prometheus.GaugeValue,
-		float64(ap.ServiceUsecPerMirroredWriteOp),
+		ap.ServiceUsecPerMirroredWriteOp,
 		"service_usec_per_mirrored_write_op",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.LatencyDesc,
 		prometheus.GaugeValue,
-		float64(ap.ServiceUsecPerReadOp),
+		ap.ServiceUsecPerReadOp,
 		"service_usec_per_read_op",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.LatencyDesc,
 		prometheus.GaugeValue,
-		float64(ap.ServiceUsecPerWriteOp),
+		ap.ServiceUsecPerWriteOp,
 		"service_usec_per_write_op",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.LatencyDesc,
 		prometheus.GaugeValue,
-		float64(ap.UsecPerMirroredWriteOp),
+		ap.UsecPerMirroredWriteOp,
 		"usec_per_mirrored_write_op",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.LatencyDesc,
 		prometheus.GaugeValue,
-		float64(ap.UsecPerReadOp),
+		ap.UsecPerReadOp,
 		"usec_per_read_op",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.LatencyDesc,
 		prometheus.GaugeValue,
-		float64(ap.UsecPerWriteOp),
+		ap.UsecPerWriteOp,
 		"usec_per_write_op",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.LatencyDesc,
 		prometheus.GaugeValue,
-		float64(ap.ServiceUsecPerReadOpCacheReduction),
+		ap.ServiceUsecPerReadOpCacheReduction,
 		"service_usec_per_read_op_cache_reduction",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.LatencyDesc,
 		prometheus.GaugeValue,
-		float64(ap.LocalQueueUsecPerOp),
+		ap.LocalQueueUsecPerOp,
 		"local_queue_usec_per_op",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.LatencyDesc,
 		prometheus.GaugeValue,
-		float64(ap.UsecPerOtherOp),
+		ap.UsecPerOtherOp,
 		"usec_per_other_op",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.BandwidthDesc,
 		prometheus.GaugeValue,
-		float64(ap.MirroredWriteBytesPerSec),
+		ap.MirroredWriteBytesPerSec,
 		"mirrored_write_bytes_per_sec",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.BandwidthDesc,
 		prometheus.GaugeValue,
-		float64(ap.ReadBytesPerSec),
+		ap.ReadBytesPerSec,
 		"read_bytes_per_sec",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.BandwidthDesc,
 		prometheus.GaugeValue,
-		float64(ap.WriteBytesPerSec),
+		ap.WriteBytesPerSec,
 		"write_bytes_per_sec",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.ThroughputDesc,
 		prometheus.GaugeValue,
-		float64(ap.MirroredWritesPerSec),
+		ap.MirroredWritesPerSec,
 		"mirrored_writes_per_sec",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.ThroughputDesc,
 		prometheus.GaugeValue,
-		float64(ap.ReadsPerSec),
+		ap.ReadsPerSec,
 		"reads_per_sec",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.ThroughputDesc,
 		prometheus.GaugeValue,
-		float64(ap.WritesPerSec),
+		ap.WritesPerSec,
 		"writes_per_sec",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.ThroughputDesc,
 		prometheus.GaugeValue,
-		float64(ap.OthersPerSec),
+		ap.OthersPerSec,
 		"others_per_sec",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.AverageSizeDesc,
 		prometheus.GaugeValue,
-		float64(ap.BytesPerMirroredWrite),
+		ap.BytesPerMirroredWrite,
 		"bytes_per_mirrored_write",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.AverageSizeDesc,
 		prometheus.GaugeValue,
-		float64(ap.BytesPerOp),
+		ap.BytesPerOp,
 		"bytes_per_op",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.AverageSizeDesc,
 		prometheus.GaugeValue,
-		float64(ap.BytesPerRead),
+		ap.BytesPerRead,
 		"bytes_per_read",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.AverageSizeDesc,
 		prometheus.GaugeValue,
-		float64(ap.BytesPerWrite),
+		ap.BytesPerWrite,
 		"bytes_per_write",
 	)
 	ch <- prometheus.MustNewConstMetric(
 		c.QueueDepthDesc,
 		prometheus.GaugeValue,
-		float64(ap.QueueDepth),
+		ap.QueueDepth,
 	)
 }
 
 func NewArraysPerformanceCollector(fa *client.FAClient) *ArrayPerformanceCollector {
 	return &ArrayPerformanceCollector{
-                LatencyDesc: prometheus.NewDesc(
-                        "purefa_array_performance_latency_usec",
-                        "FlashArray array latency",
-                        []string{"dimension"},
-                        prometheus.Labels{},
-                ),
-                ThroughputDesc: prometheus.NewDesc(
-                        "purefa_array_performance_throughput_iops",
-                        "FlashArray array throughput",
-                        []string{"dimension"},
-                        prometheus.Labels{},
-                ),
-                BandwidthDesc: prometheus.NewDesc(
-                        "purefa_array_performance_bandwidth_bytes",
-                        "FlashArray array throughput",
-                        []string{"dimension"},
-                        prometheus.Labels{},
-                ),
-                AverageSizeDesc: prometheus.NewDesc(
-                        "purefa_array_performance_average_bytes",
-                        "FlashArray array average operations size",
-                        []string{"dimension"},
-                        prometheus.Labels{},
-                ),
-                QueueDepthDesc: prometheus.NewDesc(
-                        "purefa_array_performance_queue_depth_ops",
-                        "FlashArray array queue depth size",
-                        []string{},
-                        prometheus.Labels{},
-                ),
+		LatencyDesc: prometheus.NewDesc(
+			"purefa_array_performance_latency_usec",
+			"FlashArray array latency",
+			[]string{"dimension"},
+			prometheus.Labels{},
+		),
+		ThroughputDesc: prometheus.NewDesc(
+			"purefa_array_performance_throughput_iops",
+			"FlashArray array throughput",
+			[]string{"dimension"},
+			prometheus.Labels{},
+		),
+		BandwidthDesc: prometheus.NewDesc(
+			"purefa_array_performance_bandwidth_bytes",
+			"FlashArray array throughput",
+			[]string{"dimension"},
+			prometheus.Labels{},
+		),
+		AverageSizeDesc: prometheus.NewDesc(
+			"purefa_array_performance_average_bytes",
+			"FlashArray array average operations size",
+			[]string{"dimension"},
+			prometheus.Labels{},
+		),
+		QueueDepthDesc: prometheus.NewDesc(
+			"purefa_array_performance_queue_depth_ops",
+			"FlashArray array queue depth size",
+			[]string{},
+			prometheus.Labels{},
+		),
 		Client: fa,
 	}
 }
