+++ internal/rest-client/directories_performance.go
@@ -4,18 +4,18 @@ package client
 type DirectoryPerformance struct {
 	Id      string  `json:"id"`
 	Name      string  `json:"name"`
-	BytesPerOp      int  `json:"bytes_per_op"`
-	BytesPerRead      int  `json:"bytes_per_read"`
-	BytesPerWrite      int  `json:"bytes_per_write"`
-	OthersPerSec      int  `json:"others_per_sec"`
-	ReadBytesPerSec      int  `json:"read_bytes_per_sec"`
-	ReadsPerSec      int  `json:"reads_per_sec"`
+	BytesPerOp      float64  `json:"bytes_per_op"`
+	BytesPerRead      float64  `json:"bytes_per_read"`
+	BytesPerWrite      float64  `json:"bytes_per_write"`
+	OthersPerSec      float64  `json:"others_per_sec"`
+	ReadBytesPerSec      float64  `json:"read_bytes_per_sec"`
+	ReadsPerSec      float64  `json:"reads_per_sec"`
 	Time      int  `json:"time"`
-	UsecPerOtherOp      int  `json:"usec_per_other_op"`
-	UsecPerReadOp      int  `json:"usec_per_read_op"`
-	UsecPerWriteOp      int  `json:"usec_per_write_op"`
-	WriteBytesPerSec      int  `json:"write_bytes_per_sec"`
-	WritesPerSec      int  `json:"writes_per_sec"`
+	UsecPerOtherOp      float64  `json:"usec_per_other_op"`
+	UsecPerReadOp      float64  `json:"usec_per_read_op"`
+	UsecPerWriteOp      float64  `json:"usec_per_write_op"`
+	WriteBytesPerSec      float64  `json:"write_bytes_per_sec"`
+	WritesPerSec      float64  `json:"writes_per_sec"`
 }
 
 type DirectoriesPerformanceList struct {
